+++ qa/rpc-tests/test_framework/zip317.py
@@ -24,9 +24,9 @@ GRACE_ACTIONS = 2
 # https://zips.z.cash/zip-0317#recommended-algorithm-for-block-template-construction
 WEIGHT_RATIO_CAP = 4
 
-# Limits the number of unpaid actions in a block. See
+# Default limit on the number of unpaid actions in a block. See
 # https://zips.z.cash/zip-0317#recommended-algorithm-for-block-template-construction
-BLOCK_UNPAID_ACTION_LIMIT = 50
+DEFAULT_BLOCK_UNPAID_ACTION_LIMIT = 50
 
 # The zcashd RPC sentinel value to indicate the conventional_fee when a positional argument is
 # required.